requirements.txt
absl-py==0.3.0
astor==0.7.1
cycler==0.10.0
gast==0.2.0
grpcio==1.14.1
h5py==2.8.0
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.3
numpy==1.14.5
opencv-python==3.4.2.17
Pillow==5.2.0
pkg-resources==0.0.0
protobuf==3.6.0
pyparsing==2.2.0
python-dateutil==2.7.3
python-mnist==0.6
pytz==2018.5
PyYAML==3.13
scipy==1.1.0
six==1.11.0
tensorboard==1.10.0
#tensorflow-gpu==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
Werkzeug==0.14.1
